TaskProject/requirements.txt
django==5.0.4
psycopg2==2.9.9
djangorestframework
